Split up tests. Appease Scrutinizer
