+++ .gitignore
@@ -2,3 +2,4 @@ phpunit.xml
 composer.lock
 composer.phar
 vendor
+.idea
